requirements.txt
ddgs
requests
Pillow
streamlit
pyperclip
re
urllib.parse
io
time
functools
wraps
urllib.parse